refactor(ActionType): move ActionType into store
